Add normal distribution to SampledInt adjacency
For a `SampledInt`, we want to be able to choose how we use the `adjacency` parameter.

Setting the new `uniform` argument to `True` chooses a random integer between `1` and `adjacency` for which to move the residue. 

Setting `uniform` to `False` utilizes the `adjacency` value as the standard deviation of a normal distribution centered at zero. The magnitude of the move is the integer value higher than the draw of the move.  Thus, the move set will be biased towards small moves, with a non-zero chance of selecting moves up to 3 times the value of `adjacency`.